+++ uikit/src/main/java/com/sendbird/uikit/fragments/ChatNotificationChannelFragment.java
@@ -349,6 +349,20 @@ public class ChatNotificationChannelFragment extends BaseModuleFragment<ChatNoti
         }
 
         /**
+         * Sets whether the left button of the header is used.
+         *
+         * @param useHeaderLeftButton <code>true</code> if the left button of the header is used,
+         *                            <code>false</code> otherwise.
+         * @return This Builder object to allow for chaining of calls to set methods.
+         * @since 3.5.2
+         */
+        @NonNull
+        public Builder setUseHeaderLeftButton(boolean useHeaderLeftButton) {
+            bundle.putBoolean(StringSet.KEY_USE_HEADER_LEFT_BUTTON, useHeaderLeftButton);
+            return this;
+        }
+
+        /**
          * Sets the custom fragment. It must inherit {@link ChatNotificationChannelFragment}.
          *
          * @param fragment custom fragment.